Un solo conflicto, en el constructor de `WebView`: ZAB-68 pasa a `GLRenderer`
el callback de repintado que restaura el contexto perdido, y ZAB-72 toca las
dos líneas de al lado (`onDiagnostic` y `parseColor` como método de la vista).
Los dos cambios son independientes; se quedan los dos.

El resto lo resolvió git: el `alive()` con el que ZAB-68 guarda los mutadores
del handle convive con el getter de `viewIds` — que es un LECTOR y por tanto
sigue respondiendo en una vista disposed, la regla que ZAB-68 fija para
`snapshot` y `stats`.

Verificado tras el merge: build, typecheck, 959 tests y lint en verde.

+++ packages/renderer-web/src/tessellator.ts
@@ -30,7 +30,8 @@ interface BatchStore {
   vertices: Float32Array;
   /** Floats written so far (8 per vertex). */
   floats: number;
-  indices: Uint16Array;
+  /** 32-bit: a batch past 65.535 vertices would wrap 16-bit indices (ZAB-68). */
+  indices: Uint32Array;
   /** Indices written so far. */
   count: number;
 }
@@ -109,7 +110,7 @@ export class GeometryBuilder {
       clip,
       vertices: new Float32Array(1024),
       floats: 0,
-      indices: new Uint16Array(512),
+      indices: new Uint32Array(512),
       count: 0,
     };
   }
@@ -441,10 +442,10 @@ function growFloats(array: Float32Array, needed: number): Float32Array {
   return next;
 }
 
-function growIndices(array: Uint16Array, needed: number): Uint16Array {
+function growIndices(array: Uint32Array, needed: number): Uint32Array {
   let capacity = array.length * 2;
   while (capacity < needed) capacity *= 2;
-  const next = new Uint16Array(capacity);
+  const next = new Uint32Array(capacity);
   next.set(array);
   return next;
 }
